Use the message queue again. To recap, we want it for two reasons:
* We can't look up the unique ID of an endpoint while inside the Monitor() function.
(Our workaround earlier was to create a MIDIClient, use it to look up endpoints when they changed, and construct a mapping from MIDIEndpointRef to unique ID. Complicated, but it works.)

* Doing too much while in the Monitor() routine causes all MIDI transmission to bog down.
Apparently just a CFMessagePortSendRequest() is enough to cause problems.

The message queue now allows any CFType to be kept in it--very clean. And the driver is now a lot simpler.