Changed takePendingPacketList into a block, called on a configurable queue.
Previously just used performSelectorOnMain…. Now MIDI reading can continue on a dispatch queue set by the new readQueue property of SMInputStream. Allows the issues where the main thread may be blocked (for example, by minimising a window or animating a sheet) to be fixed to continue reading MIDI in the background.

May have to allocate more memory on the time constrained Core MIDI thread, when it might have to copy the block.