Tweak autosave and close behavior a little more.
If full autosave is turned off in the system prefs, then don't dirty the document when
sections are hidden and shown. Also, in that case, pay attention to the "ask to
keep changes when closing documents" setting.

If full autosave is turned on, then we can dirty the document
all the time with impunity, and we don't need to pay attention to the "ask to keep
changes when closing documents" setting.

Added text to the prefs panel trying to explain this.