Solve a bad bug. When a new destination appeared and we created a new window to spy on it, and also "act as a destination" was set for new documents, we would keep creating new windows over and over.
In the new document, we were creating a virtual destination, then removing it immediately afterwards. We immediately removed the Destination from the object list. However, CoreMIDI still had enqueued notifications about the virtual destination. When we got the ObjectAdded notification for it, we created a new Destination since the original one was gone.  And, we couldn't tell it was our own, so we made a new window for it.  Repeat.

To fix: don't remove the Destination right away, and instead wait for CoreMIDI to notify us that it has been removed. There's no actual need to make it go away early.

(Do the same for virtual sources, even though nothing in either app uses that functionality.)