Fix #96: MIDI Monitor in full screen: hiding/showing Sources or Filters causes broken layout.
DisclosableView expects to be able to resize the window taller or shorter, but of course that won't work in full screen. Instead, resize the views appropriately, with animation.

(Note that if one of the moving views has a keyboard focus ring on it -- e.g. the "number of messages" text field -- then the focus ring jumps into place instead of animating smoothly with its view. I couldn't find an easy way around that. No such bug when you resize the window, of course...)