Fix problem resolving Sparkle package version
The problem: this app needs Sparkle 2, but the Hex Fiend submodule's project had a reference to Sparkle 1. Even though that was in a target that we didn't rely on, the Swift package manager was still unhappy with it.

Fix by pointing the Hex Fiend submodule at my fork of it, with a change to make it get Sparkle 2.

Also explicitly import Sparkle from Swift code, and don't put it in the bridging header.