Fix hang when opening SEL if library entry points to unmounted network volume
If a library entry's alias pointed at an unmounted network volume, apparently `URL(resolvingBookmarkData:)` sometimes hangs or takes a long time. If we are using the option `.withoutUI`, also include `.withoutMounting` so we don't try to automatically mount the volume.

See https://stackoverflow.com/a/2235076/1218876 which explains that `.withoutUI` just prevents "the UI potentially involved in bookmark/alias resolution ... displayed if the target is on an unmounted fileserver which requires authentication to connect."