Fix a warning (will be an error in Swift 6) in InputStream's midiRead function
There's a warning about `Optional<UnsafeMutableRawPointer>` not conforming to `Sendable`. This doesn't appear easy to fix. Sigh.

Fix: We had a `var readQueue: DispatchQueue` that was always `DispatchQueue.main` in practice. If we just use `DispatchQueue.main` literally in the code, then that apparently triggers a special case in the compiler that inserts `@_unsafeMainActor` which causes the warning to go away. Hurray, I guess.

https://oleb.net/2024/dispatchqueue-mainactor/