Update HexFiend
I'm still using my own fork of it, because their project still references a specific version of Sparkle which is very old and not compatible with ours.

It looks like they have made some progress in making it work as a Swift package, but it isn't there yet. They don't seem to have published a `Package.swift` file. Maybe someday.